pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Pavel Bláha <pwhprown@seznam.cz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.2"
djangorestframework = "^3.16.0"
django-cors-headers = "^4.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
